PythonProject/requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
APScheduler==3.10.4
openpyxl==3.15.0
tzlocal==5.2
